transaction card slight remodel and now accepts new input type of currency.

+++ src/components/TransactionsHistory/TransactionCard.tsx
@@ -1,4 +1,4 @@
-import { useContext, useState } from "react";
+import { MouseEventHandler, useContext, useState } from "react";
 import { GlobalDataContext, GlobalDataContextType } from "../../Data/Contexts/GlobalDataContext";
 import { Transaction } from "../../Data/Types/Transaction";
 import { Database, DatabaseStores } from "../../Data/Database";
@@ -7,6 +7,7 @@ import { InputText } from "primereact/inputtext";
 import { Dropdown } from "primereact/dropdown";
 import { InputNumber } from "primereact/inputnumber";
 import { Calendar } from "primereact/calendar";
+import CurrencyDropDown from "../form/CurrencyDropDown";
 
 
 const TransactionCard: React.FC<{ transaction: Transaction }> = (props): JSX.Element => {
@@ -14,10 +15,9 @@ const TransactionCard: React.FC<{ transaction: Transaction }> = (props): JSX.Ele
    const [transactionData, setTransactionData] = useState(props.transaction);
    const context = useContext(GlobalDataContext) as GlobalDataContextType;
 
-   const ValidCurrencies = [{"ABC":"ABC","DEF":"DEF"}]
-
    const backgroundColor = transactionData.type == "Expense" ? "bg-red-200" : "bg-green-200"
-   const textColor = transactionData.type == "Expense" ? "red" : "green"
+   const borderColor = transactionData.type == "Expense" ? "border-red-300" : "border-green-300"
+   //const textColor = transactionData.type == "Expense" ? "red" : "green"
    const icon = transactionData.type == "Expense" ? "pi pi-send" : "pi pi-dollar"
    const transactionDataTime = new Date(transactionData.dateTime)
 
@@ -37,81 +37,88 @@ const TransactionCard: React.FC<{ transaction: Transaction }> = (props): JSX.Ele
       setBusy(false);
    }
 
-   return (
-      <article className="flex gap-1 flex-wrap max-w-29rem surface-ground p-2 shadow-3 border-round-2xl">
+   const confirmAction = async () => {
+      
+      const confirmation = confirm("Are you sure you want to delete this transaction?");
+
+      if (confirmation) {
+         setBusy(true)
+         const db = await Database.getDB();
+         const data = context.data.Finances
+
+
+         db.delete(DatabaseStores.Finances, transactionData.id)
+         let indexToDelete = data.findIndex(t => t.id == transactionData.id)
+         context.UpdateData(prevData => ({
+            ...prevData,
+            Finances: data.filter((_, index) => index !== indexToDelete)
+         }));
          
-         <header className="text-justify flex ">
-            <i className={` ${icon} ${backgroundColor} align-content-center border-round-3xl text-4xl p-2 `} />
+         setBusy(false)
+         confirm("Transaction deleted!")
+      }
+   }
+
+   return (
+      <article className={` border-1 ${borderColor} flex gap-1 flex-wrap max-w-30rem surface-ground p-2 shadow-3 border-round-2xl `}>
+
+         <header className={`  text-justify flex justify-content-between w-full `}>
+            <i className={` ${icon} ${backgroundColor}  align-content-center border-round-3xl text-4xl p-2 `} />
+
+            <span>
+               <Button className="align-self-end" icon="pi pi-trash" severity="danger" disabled={busy} outlined
+                  onClick={confirmAction} />
+            </span>
          </header>
-         <section className=" text-justify px-1  sm:w-22rem sm:max-w-22rem">
-            <InputText value={transactionData.title} 
+         <section className=" text-justify px-1 flex flex-column flex-grow-1 gap-1 ">
+            <InputText value={transactionData.title}
                className="border-none text-xl surface-ground"
                onChange={(e) => {
                   setTransactionData(prevData => ({ ...prevData, title: e.target.value }));
                }}
                onBlur={HandleBlur}
             />
-            <div>
-               <Dropdown  options={Object.keys(ValidCurrencies)}
-                  filter value={transactionData.ammountType}
-                  className="hidden-input"
+            <div className="flex gap-1 justify-content-start align-items-center">
+               <CurrencyDropDown className="hidden-input "
+                  loading={context.data.FetchingCurrencies}
+                  options={context.data.ValidCurrencies}
+                  value={transactionData.ammountType}
                   onChange={(e) => {
                      setTransactionData(prevData => ({ ...prevData, ammountType: e.value }));
                   }}
                   onBlur={HandleBlur}
                />
-               <InputNumber className={` hidden-input m-1 p-0 text-lg ${textColor} `}
+
+               <InputNumber className=" hidden-input "
                   defaultValue={0}
                   value={transactionData.amount} minFractionDigits={0} maxFractionDigits={20}
                   onBlur={HandleBlur}
                   onChange={(e) => {
-                     setBusy(true)
-                     setTransactionData(prevData => ({ 
-                        ...prevData, 
-                        type: transactionData.amount < 0 ? "Expense" : "Income",
-                        amount: e.value as number }));
-                     setBusy(false)
+                     setBusy(true);
+                     setTransactionData(prevData => ({
+                        ...prevData,
+                        amount: e.value == null ? 0 : e.value,
+                        type: e.value == null ? "Expense" : e.value <= 0 ? "Expense" : "Income",
+                     }));
+                     setBusy(false);
                   }}
                />
             </div>
 
          </section>
-         <aside className=" flex sm:justify-content-start justify-content-end w-full sm:w-min sm:flex-column gap-1">
-            <Button icon="pi pi-trash" severity="danger" disabled={busy} outlined
-               onClick={async () => {
-                  setBusy(true)
-                  const db = await Database.getDB();
-                  const data = context.data.Finances
 
-
-                  db.delete(DatabaseStores.Finances, transactionData.id)
-                  let indexToDelete = data.findIndex(t => t.id == transactionData.id)
-                  context.UpdateData(prevData => ({
-                     ...prevData,
-                     Finances: data.filter((_, index) => index !== indexToDelete)
-                  }));
-
-                  setBusy(false)
-               }}>
-            </Button>
-         </aside>
-         <footer className="w-full flex">
+         <footer className="w-full flex justify-content-between">
             <div className="flex w-full justify-content-end pr-2 text-color-secondary">
                <span>
                   <Calendar touchUI={context.data.User.IsMobile}
                      onChange={(e) => {
                         setBusy(true);
-                        console.log((e.value))
-
                         setTransactionData(prevData => ({ ...prevData, dateTime: (e.value as Date).getTime() }));
-
                         setBusy(false);
                      }}
                      onBlur={HandleBlur}
                      className="hidden-input" showTime
-                     value={transactionDataTime}
-
-                  />
+                     value={transactionDataTime} />
                </span>
             </div>
          </footer>
